port(linux): Batch 1 - widen WindowMsgData, fix pointer-in-int sibling casts
GameWindow.h: change WindowMsgData from `typedef UnsignedInt` to
`typedef uintptr_t`. This type carries raw pointers through the GUI
callback/message-dispatch system (GameWinInputFunc/GameWinSystemFunc,
Gadget.h's callback declarations) via direct cast - safe to widen per
Draft 11's verification: dispatch is synchronous (no message queue),
never stored in a struct, never serialized (no xfer/Snapshot/CRC
contact), and a zero-effect no-op on the only shipping 32-bit targets
(uintptr_t == UnsignedInt there). Single typedef change resolved 190
of 446 WSL2 build errors - the same one-header-many-errors leverage
KeyDefs.h/GameState.h had.

Fixed the sibling pattern the typedef alone can't reach: small Int/
UnsignedInt/enum values encoded into void* callback-data slots
(GadgetListBoxGetItemData/GadgetComboBoxGetItemData/GadgetButtonGetData/
winGetUserData's return value, and userData/param arguments in INI
parse callbacks and GameSpy thread callbacks), cast back with a
literal 32-bit target type. These are value-truncation-safe (the
stored value was always an Int) but GCC correctly rejects the
pointer-to-smaller-int cast; fixed with a `(Type)(uintptr_t)expr`
double-cast at ~45 call sites across both trees. One site
(GadgetListBox.cpp's GLM_GET_SELECTION multi-select branch) actually
stores a real pointer through this slot, not an encoded int - fixed
by writing through `Int**` instead of truncating, which is more
correct than the mechanical idiom (a plain (Int) cast there would
have written only 4 of 8 bytes into the caller's pointer variable,
a real 64-bit bug the compile error was masking).

Verified: WSL2 g_gameenginedevice build dropped from 446 to 194
unique error lines. Windows g_gameenginedevice/z_gameenginedevice
build with 0 errors.

+++ Generals/Code/GameEngine/Source/Common/Thing/ThingTemplate.cpp
@@ -452,7 +452,7 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 {
 	ThingTemplate* self = (ThingTemplate*)instance;
 	ModuleInfo* mi = (ModuleInfo*)store;
-	ModuleType type = (ModuleType)(UnsignedInt)userData;
+	ModuleType type = (ModuleType)(UnsignedInt)(uintptr_t)userData;
 	const char* token = ini->getNextToken();
 	AsciiString tokenStr = token;
 
@@ -553,7 +553,7 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 //-------------------------------------------------------------------------------------------------
 void ThingTemplate::parseIntList(INI* ini, void *instance, void* store, const void* userData)
 {
-	Int numberEntries = (Int)userData;
+	Int numberEntries = (Int)(uintptr_t)userData;
 	Int *intList = (Int*)store;
 
 	for( Int intIndex = 0; intIndex < numberEntries; intIndex ++ )
